drivers: refactor: Make MU implementation generic
Messaging Unit can be found on all i.MX8 platforms so move
the implementation into a common place so that it can be
used by everyone.

Users need to define their own specific MU_BASE address.
